feat: add bid history in bid page[#8]
